+++ src/toisto/model/topic/topic.py
@@ -1,0 +1,13 @@
+"""Topic model."""
+
+from dataclasses import dataclass
+
+from ..language.concept import ConceptId
+
+
+@dataclass(frozen=True)
+class Topic:
+    """Topic model."""
+
+    name: str
+    concepts: tuple[ConceptId, ...]