+++ tests/concepts/test_concepts.py
@@ -5,7 +5,7 @@ from typing import cast
 
 from toisto.metadata import CONCEPT_JSON_FILES
 from toisto.model.language.concept import Concept, ConceptId
-from toisto.model.language.concept_factory import ConceptDict, Topic, create_concept
+from toisto.model.language.concept_factory import ConceptDict, create_concept
 from toisto.persistence.concepts import ConceptIdRegistry, load_concepts
 
 from ..base import ToistoTestCase
@@ -16,7 +16,7 @@ class ConceptsTest(ToistoTestCase):
 
     def setUp(self) -> None:
         """Override to set up test fixtures."""
-        self.concept = create_concept(ConceptId("welcome"), cast(ConceptDict, {}), Topic("topic"))
+        self.concept = create_concept(ConceptId("welcome"), cast(ConceptDict, {}))
         argument_parser = ArgumentParser()
         self.concepts = load_concepts(CONCEPT_JSON_FILES, ConceptIdRegistry(argument_parser), argument_parser)